requirements.txt
asgiref==3.8.1
Babel==2.15.0
Django==4.2.3
django-countries==7.6.1
django-phonenumber-field==7.3.0
phonenumbers==8.13.39
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
